KUDU-162 - Remove the extra undo arena and instead use the CompactionInput ones
Some testing and profiling seem to indicate that while the release and
debug builds are the same as pre-undos the asan build took a big hit
and that the extra undo arena might be the culprit.

This adds a PreparedBlockArena() method to CompactionInput that allows
to reuse the input's arena instead of creating one and continuously resetting
it.

Before this patch jenkins still timeouts sometimes (build 2034) on the mt-tablet-test
on asan after this patch the mt-tablet-test on asan goes back to the pre-undos times
of < 200secs (build 2035).

Change-Id: I56d9bb1a87a41ee183c226c2d48e15d5c4cc10d2
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1883
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>