Add static thread local implementations
Two thread local implementations, one for block-scoped thread locals
(inside functions) and one for static member variables of classes.

The implementation / interface in both cases is based on macros.

For the block-scoped implementation, usage is similar to a C++11 static
thread_local. The BLOCK_STATIC_THREAD_LOCAL macro defines and
initializes a thread local pointer whose implementation is allocated on
the heap. The constructor is lazily called whenever a new thread enters
the block, and the object's destructor is called when the thread's
Runnable completes or the thread exits.

For the static class member implementation, macros must be called to
declare, define, and initialize the thread local. The initialization
macro, INIT_STATIC_THREAD_LOCAL, must be called by any thread that may
access the thread local object before that access occurs.

Also added unit tests for each implementation.

Change-Id: Ia674ab7826d97b90ce76e226e1f47c3fbd8c5004
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1908
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>