KUDU-171. Fix tracing-related crashes in QueueInboundCall
This fixes a race where a call could be enqueued, and then processed so quickly
that the "Inserted onto call queue" trace message was added after it had
already been responded to and deleted.

The fix changes the trace message to be added _before_ it inserts onto the call
queue, when we can guarantee the pointer is still valid.

This is a race which is hard to test, but adding a 100ms sleep after the
QUEUE_SUCCESS response caused it reproduce reliably before. With the reordered
trace, the same sleep causes no problem.

Change-Id: I513d29871fed87896ee26f950d572d753a5aeb13
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1935
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>