KUDU-28. Fix failure in mt-tablet-test TestBlowOutServiceQueue
This fixes a test error where the test was too dependent on timing.
The test used to sleep for a fixed 100ms assuming that was enough time
for an RPC call to get processed. Generally that was the case, but on
an overloaded machine it could fail. By changing the long sleep to only
1ms it would fail nearly every time.

The fix is to use a latch to explicitly coordinate the test timing.

Change-Id: I05911c0371905f9758c55ea932493ea0ca45e47b
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1936
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>