HybridClock - Bug Fix: Never trim the waited time
Turns out that if the clock's error is reset _after_ we've obtained a
timestamp there was a corner case where we were retuning from WaitUntilAfter()
before the time passed as argument. In particular we were trimming the time
to min(the_latest, now_latest) and we can't ever do that. Even if we know that
'then_lastest' can't possibly be so much in the future we can't change a past
clock read (which might have been used as a timestamp for a write)
and should therefore wait until it has passed as usual.

This also fixes the test to account for the same situation.

Change-Id: I61761f0ce3048c7b3a55da38ca71bfa950f04620
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1997
Tested-by: David Alves <david.alves@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>