Minor fixes to tablet_server-test: - Disable the default fsync - Increase out-of-sync clock tolerance
In tests we sometimes update the clock to 2/3 times the error. Since the
error is frequently 1.5 seconds this sometimes brings the clock
dangerously close to the 5 sec out-of-sync limit. This patch increased
that limit to 10 sec, so that we are safe.

This also turns off fsync in the tablet server test. This was there so
that we could get history on the latencies for the write lantency bench.
but it has too great an impact on the test. Besides it can be turned on
through the flag.

Change-Id: I01e1b1d841f3428d7782088e97201569246f2457
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1999
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>