KUDU-161. Fix TS SEGV on shutdown due to OpIdAnchorRegistry destruction
Previously, the OpIdAnchorRegistry was solely owned by the TabletPeer. When a
TS shuts down, it destroys the TabletPeer, but scanners (or flush threads) may
continue to hold on to memory stores. Those memory stores then hold a reference
to the registry (through their anchor object). So, if the shutdown order isn't
carefuly orchestrated, we get a SEGV when they try to unregister themselves.

This simple patch makes OpIdAnchorRegistry ref-counted, and has each of the
anchors hold a reference.

Change-Id: I7dd96db6a9f0be9af4792ef4fd9e5620ea44044e
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2010
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>