KUDU-182 - Move the log into local scope to avoid a race
This solves a test only problem where the log would SIGSEGV
on the peer callback. The log is a member of the test class
and the peer is a local test method var, so the peer would
get destroyed before the log because it has local scope and
when the log would finally drain there'd be a SIGSEV because
the peer was already destroyed.

Solved this by also making the log a local method var.

Change-Id: I21c1bc92a3d4f99b5bd10c132a6f3f323d081eff
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2013
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>