KUDU-173 Wait for pending tasks to complete on TS shutdown
On TS/TabletPeer shutdown we don't wait/abort pending tasks.
So we end up in situation where a pending task is trying to write
on a closed log or similar.

We store the active set of Transactions so that we may wait for them to
become inactive before Shutdown. We also make Transactions refcounted so
that other subsystems may prevent immediate deletion of the objects when
they are removed from the active set.

Change-Id: Icc650e00c6ae5528444b31c4ae53efc0b8bc313e
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2018
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>