build: add a python script to find flaky tests
This is a simple script which uses the Jenkins API to find any tests
which have recently failed. I plan to set up a new Jenkins job which
uses this script and then re-runs only those tests in some sort of loop.

My assumption is that tests which have recently been flaky are likely to
continue to be flaky, so concentrating some extra runtime on those tests
is a good idea, even when we think we've fixed the bug.

Tested manually using -j kudu-test and '-j kudu-test -j kudu-gerrit'

Change-Id: I98f39f0f02f3438aaca22513de7e97fdfeedb87d
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2041
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>