KUDU-192 - Declare the Peer var before the Log var to insure correct destruction order
Couldn't repro until I added a sleep to Log::Close() but could repro easily after that.
Prob is similar to KUDU-182 (peer would get destroyed before the log) but the window
for it to happen would be smaller since both vars were local now. The specific problem
is local var destruction order, and fix is to declare the peer before the log.

Change-Id: I75c46df86ae2b0b79a9da9aea24d907a17048f8d
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2042
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>