Dist. Consensus - Bug fix: Solve peers test race and solve bug where peers would ACK the same message twice
A consensus peer's test was failing because of a race where we would assume
that all peers had responded to previous requests while in fact that was not
the case sometimes. The peer behavior was the expected one, i.e. a peer would
find more messages in the queue and would replicate them, however the test
was not expecting this.

This solves this by adding a WaitAllReplicated() method to test operation status
so that we can have a barrier and wait until all peers have replicated a message
before proceeding with the test.

This also solves a queue bug, where we were taking the wrong bound on the messages
map and the same peer would ack the same message twice. Solved and added a test
for this.

Change-Id: Iabb5774af8db7894d1bbdeddf151f556ca783166
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2048
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins