[java client] Wait for the TS to connect to the master before starting the tests
Added a method that calls ListTabletServers on the master, and BaseKuduTest can
use it to make sure it has at least 1 TS before continuing. It also works if
there are more than 1.

As a nice side effect, we used to wait 1 second for the TS to come up, but if
it's ready before that then now we'll notice it.

Change-Id: Ide20b872d414fc8f1758bbc33ae592b9f5503190
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2044
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins