rpc: fix thread-safety bugs
TSAN flags an error in Connection::ToString because it accesses many internal
state fields of the connection object. Connections aren't synchronized since
they are expected to only be accessed from the reactor thread.

There was also an error where the acceptor thread accesses reactor->cur_time()
which is unsynchronized. Changed that to simply use the current time, since
this is once-per-connection and thus using the "cached" time is not a win.

Change-Id: I2a13d47a60d2eaac7d53f0032365955051c6e669
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2052
Tested-by: Michael Percy <mpercy@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>