metadata: beginnings of a convenient API for specifying atomic changes
Previously there are parts in the code which make several successive operations
to modify a RowSetMetadata. We also have some complex APIs that take several
arguments in order to atomically swap in/out metadata entries.

This adds a builder-style API to specify all of the changes in an object and
then pass that list of changes to the RowSetMetadata to be made atomically.

This is preparation work for KUDU-177.

Change-Id: Ia244e7c5ad66d3bd268b4f7093e9a0f3ec12bcef
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2070
Tested-by: jenkins
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>