Rename rw_spinlock to rw_semaphore, make rw_spinlock a wrapper
This moves the main reader-writer spinlock class to a new one, named as a
"semaphore" since it is safe to take/release on different threads, which is not
the case traditionally with a "lock". The old name is maintained as a wrapper
around the new class, with the TSAN annotations applied only there. This will
allow us to continue to get TSAN warnings if we incorrectly use the spinlock,
but not get all of the false positive TSAN warnings on cases where we expect to
acquire/release the lock cross-thread.

A later commit will change some actual usage points to rw_semaphore instead
of rw_spinlock.

Change-Id: Iac025119f0c63816c4d45c7337b859863ef5f30b
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2079
Tested-by: jenkins
Reviewed-by: Colin McCabe <cmccabe@cloudera.com>