tablet: Change component_lock_ to be an rw_semaphore
component_lock_ is picked up by transactions which start on one thread, and
later released on another. This causes a bunch of TSAN warnings that mutexes
are being released on different threads than they were acquired on -- generally
considered naughty by folks like Colin and the POSIX committee.

This switches over to the new rw_semaphore class which is allowed to be used
in this manner.

Previously, component_lock_ used some "clever" per-CPU locking to try to make
the reader path even more efficient. However, now that the lock is often
released on a different thread than it started on, it's unlikely to actually
reduce contention anymore, and instead just adds code size. It's possible this
will have some performance impact, but I expect it to be low - will watch
the benchmarks dashboard following commit.

With this patch and KUDU-82 together, I can run client-test without any
TSAN warnings.

Change-Id: I4ef1746743b833af3707672036fccc9fa7c042a4
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2083
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>