Change PartialRowsPB to RowOperationsPB, include operation type
* Renames PartialRowsPB to RowOperationsPB
* For each row, adds a byte preceding the row determining the operation type.
  Currently only INSERT is used.

This is preparatory work for changing all of the row operations
(INSERT/UPDATE/DELETE) to be in a single protobuf. The overall aim here is to
simplify some of the error reporting on the server side around handling of row
indexes in per_row_errors. It should also simplify the implementation somewhat
if all of the operation types can be handled in a single loop.

Change-Id: I0297aadc8aafa407e73f97fb90cab41abcacb4d3
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2103
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>