Dist: Consensus: Refactor consensus API, prepare for the dist impl. and add consensus fault hooks
This refactors the consensus API to be more intuitive and adds the
ReplicaTransactionFactory and ReplicaCommitContinuation interfaces which
will allow tablet peer to create replica transactions and to continue replica
transactions once the leader commit message is received.

This also adds the fault hooks for consensus. Fault hooks are not present in the
local implementation but are executed before/after each phase in dist consensus
(if registered).

Change-Id: I48792f6ef149eed4f370ab2e98b57e07d6137037
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2107
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins