KUDU-209: ts_tablet_manager-test fails sometimes in TestTabletReports
ts_tablet_manager-test seems to have a race in the TestTabletReports
test.  The code gets a tablet report from the tablet_manager and expects
that it can know what the sequence number will be.  However, this seems
fragile, since the tablet_manager may have generated a report on its own
during this time.  So let's change it to only assert that the sequence
number is going up each time we check.

Change-Id: Idd8ef421fbcfd4b84a64f939f64793319c041f9c
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2113
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins
Reviewed-by: Colin McCabe <cmccabe@cloudera.com>