Dist. Consensus - Add the remaining watermarks to Peers and PeerMessageQueue
This adds the remaining watermarks to Peers and the PeerMessageQueue. Specifically
this splits the watermark into received, replicated and committed.
The received watermark indicates the last operation the peer received. The committed
watermark indicates the last operation the peer committed (consensus replicas commit
asynchronously). The replicated watermark indicates the last operation the peer
replicated to the log. All of these are needed to correctly track remote replica
execution.

This also makes Peers always register an initial id with the queue (which might be the
reserved (0,0) id).

This also simplifies the Peer state (pending is no longer required) and adds the
ability to wait for outstanding requests on Close().

Change-Id: I6344b33f955d7ac95cc1209ba625f504fa07555b
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2119
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>