logging: fix log-related command-line args
While trying to understand KUDU-153 I noticed that the /logs endpoint thought
/tmp/UNKNOWN.INFO was my log file. After some debugging, I saw that
google::ProgramInvocationShortName was calling into gflags, not glog, as the
latter's implementation was in an internal namespace rather than "google". And
since we hadn't parsed the command line arguments yet, gflags didn't yet know
what the short name was.

The problem runs deeper: by calling InitGoogleLoggingSafe before
ParseCommandLineFlags, the FLAGS_* references in the former are guaranteed to
be unset. So to fix all this, I switched the ordering in various places and
added a check to InitGoogleLoggingSafe to ensure that gflags is initialized.

Change-Id: I14afb080db43c878513fa80f0f13380c1cc69e39
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2131
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins