[java client] Keep track of the quorum's leader, send all requests to it
This patch sends all the requests to the quorum leader instead of the first TS
in the list. It also fixes the fact that a new TabletClient was added for each
address a TS has (made no sense). It's now doing what the C++ client is doing,
that is, it picks the first address and that's it.

Change-Id: Id2a2a098b946a97dcc124b09fc6500e34bbe85b8
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2147
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>