threading (02/11): initial check-in of Impala thread wrapper
This is the Impala thread wrapper. I copied it nearly verbatim, apart from
license changes and several other changes documented at the top of each
file. For the moment it's hooked up to the build but not to any kudu code.

In general I tried to walk a fine line between preserving the original code
and adapting it to the Kudu codebase. My eventual goal is for it to be easy
for Impala and Kudu to share useful library code (such as this thread
wrapper). To that end, I tried to avoid making unnecessary changes.

One interesting semantic change is to the initialization functions. In Kudu,
a process may have multiple servers (e.g. a minicluster test), and it's
generally cleaner to handle initialization in the context of the server, not
the overall process. As such, InitThreading needs to be forgiving of
multiple calls, and StartInstrumentation should do the right thing when
called on different metric registries and webservers. This last part was
thorny, as it meant decoupling the counters from the metrics themselves.

To see only the delta from the Impala thread wrapper, check out:

  http://www.sf.cloudera.com/~adar/impala_to_kudu_import.diff

Change-Id: I853b35f892bcdb96194fbcf6fb526efe105ad306
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2256
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>