threading (04/11): fix KUDU-11 in new thread wrapper
Just to recap, this is a race wherein a thread's functor accesses thread_
before the caller is able to assign the boost thread object into it.

This wasn't something Impala cared about (perhaps because they don't run
TSAN, or because they try and share very little state between threads and
their callers). But we care about it, so I've adapted the spinlock approach
from thread_util to close the race.

Change-Id: I3b62abcf62a33be5e8efef8f66bcfd9b53e2113c
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2258
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins