threading (05/11): add thread status and don't throw boost exceptions
Kudu tries to avoid exception handling as much as possible, so instead of
throwing an exception when thread creation fails, let's just store the
result as a bad status.

A subsequent commit will make this more intuitive (i.e. instead of storing
the status, return it from a factory method).

Change-Id: Ifce4eb78017406e8d18f984a6d6bdcfdeecc45b8
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2259
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins