Clean up TxResultPB and nested PBs to unify insert/delete, simplify bootstrap
This removes a few layers of nesting in TxResultPB, and renames things to be a little
clearer:
- MutationTargetPB -> MemStoreTargetPB (since it now captures the target for insert
  or for mutate)
- TxOperationPB -> OperationResultPB (since it's the _result_ of the operation)
- MutationResultPB -> dead (since its fields were somewhat redundant)

This also cleans up bootstrap a bit so that there is now a utility function to see
if any write (MUTATE or INSERT) needs to be replayed into a given store. This
simplifies the handling of duplicated inserts, and also exposed a new potential
bug, which is captured on JIRA and a TODO in the code.

Change-Id: Ibb2e4a4a46454676799f38cf6a0470f3d95e7d13
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2165
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>