threading (08/11): expose current_thread abstraction
Just as boost exposes this_thread, it would be useful if our thread wrapper
exposed the current thread in whatever context, provided it was started
through the thread wrapper. At the very least we could use this for thread
ID comparisons (e.g. is the thread that's running the one I expect?), but it
can be used for other interesting things in the future too.

We'll implement current_thread via threadlocal storage, which means that
threads now have shared ownership: the caller and the threadlocal. This
means callers need to work with threads via scoped_refptrs instead of the
simpler gscoped_ptrs that they were already using.

As part of this commit I've also hidden the thread constructor and exposed a
factory method instead, which returns the status and stores the created
thread in a scoped_refptr.

Change-Id: I95af378c60f433d4bc48a5d66ea378adf1da7c96
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2262
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins