Make log GC logic more intuitive
Since MISSED_DELTA commits have been removed, we no longer need to
account for the possibility of duplicated OpIds, so if an anchor is
equal to the initial OpId in a log file, we can delete previous log
files, which is what one would expect. We no longer need to hold onto an
"extra" previous log file in that case. This makes the behavior and the
unit tests for log anchoring more intuitive.

Change-Id: I7dad597a8806d5d9b74328e93eb737b1aa2ebb04
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2173
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>