KUDU-227 - Read the 'before' clock time before we get 'past_max_error'
Because we were reading the 'before' clock value _after_ we had obtained
'past_max_error' there was a (small) window for failure if the test thread
got stalled in between the reads.

This patch fixes this by making the clock read the first thing that happens
in the tests therefore making sure that even if there is a thread stall
the test will no longer fail.

Change-Id: Ibcc51c4b658271dcfd28b26af5acfab58357f4ca
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2188
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>