Refactor ServicePool to contain the service queue
This change allows the ServicePool to shut down while allowing Messenger
to continue to relay client messages as long as is needed.

Also, don't allow BlockingQueue to be destroyed nonempty.
This burns us in cases where we have bare pointers on the queue, which
implies ownership of those objects by the queue. Resources can leak.

Change-Id: I4f1e9a564d74bf51d13bccdf315b413479bb101b
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2234
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>