Change the Timestamp format to uint64/fixed64
Since Timestamps are often very big the advantages of encoding them
as a var int are not obvious, plus the very specific varint 64 encoding
would have to be replicated in jave.

This changes the timestamp format to uint64/fixed64 which, given the
normal size of the timestamps would not present any encoding disadvantages
and is much easier to deal with.

Change-Id: I4673c69421b8cd1e3275275888b9a11fc160e2fc
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2245
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>