threading (9.5/11): close race where child reads tid_ before parent sets it
With the introduction of current_thread(), the set of "child visible state"
has grown to include any Thread field. In particular, tid_ is read by
several thread functors, to check that the thread running some code is
indeed the expected thread.

We need to make sure tid_ is set before StartThread() unblocks
SuperviseThread(), and that's not the case today. A simple solution is to
reverse the order of load/stores, so StartThread() only unblocks
SuperviseThread() after all private fields have been set. However, we can't
use p_c_assigned for this, as StartThread() may return before
SuperviseThread() is rescheduled, which would cause p_c_assigned to go out
of scope.

So in addition to reversing the load/store order, let's synchronize on tid_
itself instead of on p_c_assigned. Through TLS, SuperviseThread() is
guaranteed to have a reference to the Thread object, and by extension, to
tid_, so it shouldn't go out of scope.

Change-Id: Id37be33dbcf607874cca76054c2417ebc5e58668
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2273
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins