Convert a lot of "const Schema&" members to "const Schema*"
Per our style guide, we shouldn't generally have references as
data members, or pass references to functions/constructors which
end up saving a pointer to that object somewhere.

This patch converts a number of such Schema& members into pointers.
It's not complete, but these sort of patches tend to go stale quickly,
and it should be fine to incrementally improve things.

The remaining such members can be found with:
  git grep 'const [^ ]* *& *[a-z_]*_;$'

Change-Id: I4224b8db64bef8b582ec65f5c56d4e6bdc4bcac3
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3662
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins