java: automatic protoc compilation
It's pretty annoying to have to run build-proto.sh manually. Moreover, it
spits the generated code all over the source tree, and one must be careful
not to accidentally commit it.

So let's use a maven plugin to automate the usage of protoc. There isn't
an "official" version but the one maintained by David Trott appears to be
pretty good. It supports incremental building and follows maven conventions
(like putting generated code in target/generated-sources).

I spent most of the time fighting with Eclipse, and documented my findings
in the README. I also snuck in a few ancillary client-related changes.

Change-Id: I8f78987bea34f976f3ac90bd4b3b88e7d2b5268e
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2302
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: Adar Dembo <adar@cloudera.com>