Fix type of next_mrs_id_
next_mrs_id_ needs to be a 64-bit type because last_durable_mrs_id_ in
metadata.cc is 64-bit.  Also, a 32-bit type raises questions about
exhaustion.

It does not need to be atomic, since when we access it in
ReplaceMemRowSetUnlocked, we are always holding the Tablet's
component_lock_ for write.  (No atomic operations were actually done on
it previously, so that part doesn't change.)

Change-Id: I85408b024f70925f291171044bfbff3151bd576a
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2303
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>