KUDU-179: Fix locking for Tablet::schema_
This patches changes Tablet::schema_ to a shared_ptr.  Then, we set this
pointer only when holding Tablet::component_lock_ for write, and access
this pointer only when holding Tablet::component_lock_ for read.  It is
assumed that we do not modify Schema after creating it, so it is OK for
threads to take a copy of the shared_ptr, drop the lock, and access
away.  This is similar to how we handle the RowSetTree,
Tablet::rowsets_.

This avoids copying the rather large Schema object.  It means that
Tablet::schema_lock_ is no longer necessary.  It also fixes a minor race
where we were copying Tablet::schema_ in Tablet::Flush without holding
the lock.

Change-Id: I0f15bef869c09031770fdf85b6f8d2650823a779
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2317
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>