Fix json generation for metrics
This makes MetricsRegistry not start/end the metrics object itself and instead
just generate a field in that object. The object is started/ended by the /jsonmetrics
path handler and may or may not include histograms as another field in the same
object, depending on the request.

This fix was supposed to be part of http://gerrit.ent.cloudera.com:8080/#/c/2294/
but a bad rebase made it part of a throwaway patch instead.

Change-Id: I945678029a1c78669fe08e33d32f678e3ede071a
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2333
Tested-by: David Alves <david.alves@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>