Add an histogram for each write transaction external consistency mode
Histograms for external consistency modes on writes are useful for the HT paper
but can also be useful for other things, such as profiling in production which
consistency modes are most used or debugging long write times: e.g. if
writes with NO_CONSISTENCY take a long time this is probably caused by something
different than if writes with COMMIT_WAIT consistency take a long time, if the
clock error is high.

Change-Id: If8fd763606409bd3daa87f7379395858c68bde67
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2336
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>