In MRS flush, do not special-case the empty MRS
We don't need to have a special case for flushing the empty MemRowSet.
The maintenance ops scheduler should not call Flush unless there is some
benefit.  We may flush an empty memrowset sometimes when Flush is called
from AlterSchema, but schema alterations should be rare.  Having a
special case for this adds more complexity that isn't really worth it.

It's worth noting here that the current special case for flushing an
empty memrowset in Tablet::Flush (the version with arguments) does not
actually ever trigger.  The reason is because it checks for
"(input.num_rowsets() == 1 && old_ms->empty())", but in the case that
the current MemRowSet is empty, we have input.num_rowsets == 0, due to
the other special-case code in ReplaceMemRowSetUnlocked.

I added a test of modifying the schema of an empty tablet (which
therefore has an empty memrowset_) and then performing some test
operations.

Change-Id: I92f58dfc0376bc614180483bf9e680e9f1dda6c8
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2347
Tested-by: jenkins
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>