Add crcutil to thirdparty
crcutil is a highly optimized cyclic redundancy check (CRC)
implementation, which uses inline assembly to use specialized fast
hardware instructions if they are available at runtime.

While the really fast implementation is crc32c, due to Intel hardware
support via the crc32 instruction, crcutil also has hardware-assisted
software implementations of crc64 and crc128, which are likely to be
roughly an order of magnitude slower than crc32c.

Some changes needed to be made to play nicely with Kudu's uint64_t type
(from stdint.h) as well as the build scripts to generate a static
library we can include. Also, a factory method was added for crc32c, so
that the polynomial constants to get the hardware-accelerated version
do not need to be "remembered" by clients. Those changes are in the
included patch file.

Otherwise, this is stock crcutil-1.0, which has not changed since its
release in January 2011.

See https://code.google.com/p/crcutil/

Change-Id: I3aa30fc42b081e6b39c396683975e87f66f286e9
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2383
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>