Expose CRC32C as a singleton
The crcutil primary interface has a lot of input constants and params,
and the objects are relatively expensive to initialize. To avoid
restricting potential uses of the rolling CRC methods, and other various
methods exposed by the crcutil interface, we provide a simple factory
function that takes care of initializing the singleton with the params
that should be appropriate for most cases. If other use cases come up
where we need additional flexibility, it will be easy to add factory
functions for those in crc.h as needed.

The crcutil impl doesn't hold any mutable state, so in many cases
it isn't useful to have multiple instances, unless you need different
polynomials / degrees or different rolling window sizes.

The implementation is very fast on recent hardware. Output from the unit
test benchmark (slow tests off) on my laptop, which is a Core i7-3720QM
running at 2.60 GHz:

1000 runs of CRC32C on 4000000 bytes of data (total: 4000000000 bytes) in 0.15936635 seconds;
25099401.473397613 bytes per millisecond, 25 bytes per nanosecond!

Change-Id: I717660aa94e5cfbb39bc751995831226d32e8c5e
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2384
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>