KUDU-242 - Make sure generated timestamps are unique among the current in-flights
When using multiple external consistency modes at the same time there is a window
where we may assign the same timestamp to two in-flight transactions. Since
in release mode we do not check that a transaction timestamp is not already
in the in-flights sometimes we would get a CHECK failure when trying to commit
the second transaction which was assigned the same timestamp.

This patch fixes this issue by looping until the generated transaction timestamp
is unique among the current in-flights.

Change-Id: I97f28043b69453e7c7a89554ee64ee1629bd55ed
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2390
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>