Allow to only get the requested metrics and include counts/values in the same json object
Previously we would get histogram counts/values in a separate object, in the metrics
json. mpercy suggested that we should return the detail in the same metrics object,
which makes sense. This patch makes this change.

This patch also adds a way to specify which metrics are to be fetched, since tablet_servers
easily end up with 200K+ of metrics. This is done throught a "metrics" parameter that
must be a comma separated list of metrics to get back. Since some of the metrics
are tablet specific both the "metrics" parameter and the "detailed_metrics" parameter
(which allows to fetch the histogram counts/values) allows to specify a substring of the
metric name, which fetches the metrics whose names match, for all tablets.

The final url ends up being something like:

http://<host>:<port>/jsonmetricz?metrics=<m1>,<m2>&detailed_metrics=<histo1>,<histo2>

Change-Id: Ia1d6b5b6263a8e2174ce7f1ab886a3726b638ffb
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2391
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>