Make log::AsyncAppend take StatusCallback instead of FutureCallback
Throughout the code, we are using FutureCallback in a lot of places
where it would be simpler to just use StatusCallback and check
the Status argument. StatusCallback is also more convenient because
you can directly use boost::bind.

In follow-up patches I will convert other pieces to StatusCallback,
but splitting it up for easier review.

Change-Id: I96ae1119d0c011ffd496a46204bb2c9d8ae82a68
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2579
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>