KUDU-168 [java client] Handle null columns
This patch refactors the BitSet-related methods into Bytes.
For nulls it adds a BitSet in Operation for tracking the inserted
nulls and another one ine RowResult to track those coming back from
the scanners. It also adds more unit tests so that BitSets serde is
covered end to end, and in TestKuduSession it writes and reads nulls.

Change-Id: I6cf1b1560e09caa887636578c45a7251b0053983
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2410
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>