Add pstack_watcher for diagnosing test hangs.
PstackWatcher is an object which will pstack the current process and
print the results to stdout.  It does this after a certain timeout has
occured.

Sometimes pstack is installed to print the stack trace of a running
process.  (For example, on Ubuntu.)  Other times, it's gstack.
(OpenSUSE, Red Hat).  Sometimes both are installed.  We use 'which' to
check which one is installed, if any.  I added the ability to disable
stderr in Subprocess to avoid spew when using 'which' to find out this
information.

Change-Id: I2b03a49e7590cad7f8924fae9e0a25cf2ac0c82d
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2428
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins