KUDU-141: Don't let an orphan commit abort replay
Adds a test for a legitimate situation in which a COMMIT may
be orphaned. Changes the bootstrap code such that the replay does not
abort, but merely ignores the orphan.

Orphan commits that have not been flushed are still rejected with
Status::Corruption(). In the future, a checksum and other checks
may also be added to detect log corruption.

Additionally, this revision adds a unit test for normal tablet
bootstrap scenario.

Change-Id: I9c63d91ae73683554d2b1a8d718f56aeabed784d
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2429
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins