Stress test writing multiple entries to log
This revision adds a stress test for writing multiple entries
per batch to the log: mt-log-test is modified to insert multiple
entries in a batch and to verify that all of the inserted entries have
been read back.

Box-Muller based random sampling (added to util) is used to choose
batch sizes.

Change-Id: Ifee13fb5a9a9d18112767ae2ff7a295068b2d1ea
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2475
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>