client: basic multi-tablet support
This commit adds basic multi-tablet support to the C++ client. New features:
o Inserts are sent to the correct tablet and server.
o Scans can operate on multiple tablets.
o Queried tablets are cached for faster lookups later.

This is my first major commit in uncharted waters, so a thorough review is
appreciated.

The caching in the MetaCache is perhaps the most complicated piece. There's
some TSAN-compliant locking in there, but I'm not supremely confident in it.
Currently there's absolutely no cache eviction.

The order of operations for insert no longer performs an explicit refresh (to
find tablet replicas) following a lookup. In the slow lookup case this was
redundant, and if we want to handle it in the fast lookup case, we should
restrict it to when inserts fail.

Some tests were modified to account for changed semantics:
o Misuse of scanner.NextBatch(). In a multi-tablet world, nextBatch is
  called multiple times, adding to 'rows' each time. Most test code assumed
  that the contents of 'rows' was overwritten on each call.
o Scanner behavior when skipping tablets. When skipping to the next tablet,
  HasMoreRows() returns true and NextBatch()  provides no new rows. It's
  weird, but it's what the Java client does, so we'll roll with it.

Change-Id: I01d9cb66ed38ecc49ca5f6d090046a032f28378e
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2510
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>