[java client] MapReduce OutputFormat
Adds a KuduTableOutputFormat (with an empty KuduTableOutputCommitter) that can
be configured with a master address, output table, and an optional timeout.
Since creating Operations requires a KuduTable, which itself requires a
KuduClient, it was decided to have a multiton inside the output format that's
keyed by the Thread ID (it could also be a UUID). The task can reach the KuduTable
by looking up its Configuration with a special key and then passing the result to
a static method on the output format.

This is all facilitated by a util class called KuduTableMapReduceUtil which should
be used by MR developers.

Two unit tests were written (adding 9 more seconds of testing), the first one
testing the output format directly and the other one testing the output format
via a custom map-only local job.

It hasn't been tested on a real MR cluster.

Change-Id: I5d815d9473169ddcd376c6d5ea61e5a38c0efb25
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2509
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>